refactor(admin): group task monitoring module under admin/tasks
The execution-task list page, its refresh/approval logic and styles lived in separate generic folders. Move Tasks.vue, useTasks.ts and tasks.scss into one module directory.
